WIP: Fixed errors identified with PHPStan level 1 Tests to be done

+++ sources/admin.queries.php
@@ -140,27 +140,27 @@ switch ($post_type) {
         $tables = array();
         $result = DB::query('SHOW TABLES');
         foreach ($result as $row) {
-            $tables[] = $row['Tables_in_' . $database];
+            $tables[] = $row['Tables_in_' . DB_NAME];
         }
 
         //cycle through
         foreach ($tables as $table) {
             if (defined('DB_PREFIX') || substr_count($table, DB_PREFIX) > 0) {
                 // Do query
-                $result = DB::queryRaw('SELECT * FROM ' . $table);
-                DB::queryRaw(
+                $result = DB::query('SELECT * FROM ' . $table);
+                DB::query(
                     'SELECT *
                     FROM INFORMATION_SCHEMA.COLUMNS
                     WHERE table_schema = %s
                     AND table_name = %s',
-                    $database,
+                    DB_NAME,
                     $table
                 );
                 $numFields = DB::count();
 
                 // prepare a drop table
                 $return .= 'DROP TABLE ' . $table . ';';
-                $row2 = DB::queryfirstrow('SHOW CREATE TABLE ' . $table);
+                $row2 = DB::queryFirstRow('SHOW CREATE TABLE ' . $table);
                 $return .= "\n\n" . $row2['Create Table'] . ";\n\n";
 
                 //prepare all fields and datas
@@ -305,7 +305,7 @@ switch ($post_type) {
             $query .= fgets($handle, 4096);
             if (substr(rtrim($query), -1) === ';') {
                 //launch query
-                DB::queryRaw($query);
+                DB::query($query);
                 $query = '';
             }
         }
@@ -612,7 +612,7 @@ switch ($post_type) {
             break;
         }
 
-        $error = '';
+        $nb_of_items = $error = $nextStart = '';
         require_once 'main.functions.php';
 
         // prepare SK
@@ -1069,7 +1069,7 @@ switch ($post_type) {
         }
 
         // restore saltkey file
-        if (file_exists($previous_saltkey_filename)) {
+        if (isset($previous_saltkey_filename) && file_exists($previous_saltkey_filename)) {
             unlink(SECUREPATH.'/'.SECUREFILE);
             rename(
                 $previous_saltkey_filename,
@@ -1457,7 +1457,7 @@ switch ($post_type) {
         foreach ($post_list as $file) {
             if ($cpt < 5) {
                 // Get file name
-                $file_info = DB::queryfirstrow(
+                $file_info = DB::queryFirstRow(
                     'SELECT file
                     FROM ' . prefixTable('files') . '
                     WHERE id = %i',
@@ -1918,7 +1918,7 @@ switch ($post_type) {
         }
 
         // send data
-        echo '[{"result" : "' . addslashes($LANG['done']) . '" , "error" : ""}]';
+        echo '[{"result" : "' . addslashes($lang->get('done')) . '" , "error" : ""}]';
         break;
 
     case 'save_agses_options':
@@ -2033,7 +2033,7 @@ switch ($post_type) {
         }
 
         // send data
-        echo '[{"result" : "' . addslashes($LANG['done']) . '" , "error" : ""}]';
+        echo '[{"result" : "' . addslashes($lang->get('done')) . '" , "error" : ""}]';
         break;
 
     case 'save_option_change':
@@ -2492,7 +2492,7 @@ function tablesIntegrityCheck(): array
         }
 
         if ($tableFound === true) {
-            $createTable = DB::queryfirstrow("SHOW CREATE TABLE ".DB_PREFIX."$table");
+            $createTable = DB::queryFirstRow("SHOW CREATE TABLE ".DB_PREFIX."$table");
             $currentHash = hash('sha256', $createTable['Create Table']);
             if ($currentHash !== $tableHash) {
                 $tablesInError[] = DB_PREFIX.$table;
